requirements.txt
faiss-cpu>=1.8.0 #or gpu
boto3>=1.26.0
sqlalchemy>=1.4
pillow>=9.0.0
numpy>=1.24.0
fastapi>=0.95.0
uvicorn[standard]>=0.20.0
python-dotenv>=1.0.0
requests>=2.31.0
datasets
asyncpg
sentence_transformers